requirements.txt
nxt-python==3.5.1
PyBluez @ git+https://github.com/pybluez/pybluez.git@82cbba8a1ebd4c1e3442dfafd8581d58c50fa39e
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
pyusb==1.3.1
shiboken6==6.9.1